% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{summary.wyoung}
\alias{summary.wyoung}
\title{Summary method for objects of type wyoung}
\usage{
\method{summary}{wyoung}(object, digits, ...)
}
\arguments{
\item{object}{An object of type wyoung}

\item{digits}{Rounding of digits}

\item{...}{misc. function arguments}
}
\description{
Summary method for objects of type wyoung
}
